requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-tinymce==4.1.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0
requests==2.32.4
PyJWT==2.7.0
itsdangerous==2.2.0
APScheduler==3.11.0